no same username

+++ utils.py
@@ -5,8 +5,9 @@ import os
 import settings
 
 def adduser(img, username):
-    if os.path.exists("db/img/'+username+'.png"):
-        logger.info(f'"{username}" already exist. replacing image')
+    if os.path.exists(f"db/img/{username}.png"):
+        logger.info(f'"{username}" already exist')
+        return f'Username already exists'
 
     # Coppy image to image folder as username file extention
     img.save('db/img/'+username+'.png')
@@ -19,8 +20,15 @@ def adduser(img, username):
     # Create DB
     df = DeepFace.find(img_path = 'db/img/'+username+'.png', db_path = "db/img")
 
-    # Move DB to DB file
-    shutil.move('db/img/representations_vgg_face.pkl', 'db/img/db')
+    if os.path.exists(f"db/img/representations_vgg_face.pkl"):
+        logger.info("representations_vgg_face created. Moving representations_vgg_face to db/img/db")
+        # Move DB to DB file
+        shutil.move('db/img/representations_vgg_face.pkl', 'db/img/db')
+        logger.info(f'User "{username}" added')
+        return f'User "{username}" added'
+    else: 
+        logger.error(f'representations_vgg_face not created. User "{username}" not added')
+        return f'User "{username}" not added\nPlease try again'
 
 def verifyuser(img, username):
